fix: fix crashed when running in multiple flutter engine instance.
