+++ bridge/core/executing_context.h
@@ -22,7 +22,7 @@
 #include "foundation/macros.h"
 #include "foundation/ui_command_buffer.h"
 
-#include "dart_context.h"
+#include "dart_isolate_context.h"
 #include "dart_methods.h"
 #include "executing_context_data.h"
 #include "frame/dom_timer_coordinator.h"
