+++ bridge/core/html/custom/widget_element.cc
@@ -28,7 +28,7 @@ bool WidgetElement::IsValidName(const AtomicString& name) {
 }
 
 bool WidgetElement::NamedPropertyQuery(const AtomicString& key, ExceptionState& exception_state) {
-  return GetExecutingContext()->dartIsolateContext()->dartContext()->EnsureData()->HasWidgetElementShape(key);
+  return GetExecutingContext()->dartIsolateContext()->EnsureData()->HasWidgetElementShape(key);
 }
 
 void WidgetElement::NamedPropertyEnumerator(std::vector<AtomicString>& names, ExceptionState& exception_state) {
@@ -57,7 +57,7 @@ ScriptValue WidgetElement::item(const AtomicString& key, ExceptionState& excepti
     return unimplemented_properties_[key];
   }
 
-  if (!GetExecutingContext()->dartIsolateContext()->dartContext()->EnsureData()->HasWidgetElementShape(tagName())) {
+  if (!GetExecutingContext()->dartIsolateContext()->EnsureData()->HasWidgetElementShape(tagName())) {
     GetExecutingContext()->FlushUICommand();
   }
 
@@ -66,7 +66,7 @@ ScriptValue WidgetElement::item(const AtomicString& key, ExceptionState& excepti
   }
 
   auto shape =
-      GetExecutingContext()->dartIsolateContext()->dartContext()->EnsureData()->GetWidgetElementShape(tagName());
+      GetExecutingContext()->dartIsolateContext()->EnsureData()->GetWidgetElementShape(tagName());
   if (shape != nullptr) {
     if (shape->built_in_properties_.count(key) > 0) {
       return ScriptValue(ctx(), GetBindingProperty(key, exception_state));
@@ -97,12 +97,12 @@ ScriptValue WidgetElement::item(const AtomicString& key, ExceptionState& excepti
 }
 
 bool WidgetElement::SetItem(const AtomicString& key, const ScriptValue& value, ExceptionState& exception_state) {
-  if (!GetExecutingContext()->dartIsolateContext()->dartContext()->EnsureData()->HasWidgetElementShape(tagName())) {
+  if (!GetExecutingContext()->dartIsolateContext()->EnsureData()->HasWidgetElementShape(tagName())) {
     GetExecutingContext()->FlushUICommand();
   }
 
   auto shape =
-      GetExecutingContext()->dartIsolateContext()->dartContext()->EnsureData()->GetWidgetElementShape(tagName());
+      GetExecutingContext()->dartIsolateContext()->EnsureData()->GetWidgetElementShape(tagName());
   if (shape != nullptr && shape->built_in_properties_.count(key) > 0) {
     NativeValue result = SetBindingProperty(key, value.ToNative(exception_state), exception_state);
     return NativeValueConverter<NativeTypeBool>::FromNativeValue(result);
@@ -149,7 +149,7 @@ bool WidgetElement::IsAttributeDefinedInternal(const AtomicString& key) const {
 NativeValue WidgetElement::HandleSyncPropertiesAndMethodsFromDart(int32_t argc, const NativeValue* argv) {
   assert(argc == 3);
   AtomicString key = tagName();
-  assert(!GetExecutingContext()->dartIsolateContext()->dartContext()->EnsureData()->HasWidgetElementShape(key));
+  assert(!GetExecutingContext()->dartIsolateContext()->EnsureData()->HasWidgetElementShape(key));
 
   auto shape = std::make_shared<WidgetElementShape>();
 
@@ -169,7 +169,7 @@ NativeValue WidgetElement::HandleSyncPropertiesAndMethodsFromDart(int32_t argc, 
     shape->built_in_async_methods_.emplace(method);
   }
 
-  GetExecutingContext()->dartIsolateContext()->dartContext()->EnsureData()->SetWidgetElementShape(key, shape);
+  GetExecutingContext()->dartIsolateContext()->EnsureData()->SetWidgetElementShape(key, shape);
 
   return Native_NewBool(true);
 }
