+++ bridge/scripts/code_generator/templates/json_templates/make_names.cc.tpl
@@ -8,25 +8,25 @@
 namespace webf {
 namespace <%= name %> {
 
-void* names_storage[kNamesCount * ((sizeof(AtomicString) + sizeof(void *) - 1) / sizeof(void *))];
+thread_local void* names_storage[kNamesCount * ((sizeof(AtomicString) + sizeof(void *) - 1) / sizeof(void *))];
 
 <% if (deps && deps.html_attribute_names) { %>
-void* html_attribute_names_storage[kHtmlAttributeNamesCount * ((sizeof(AtomicString) + sizeof(void *) - 1) / sizeof(void *))];
+thread_local void* html_attribute_names_storage[kHtmlAttributeNamesCount * ((sizeof(AtomicString) + sizeof(void *) - 1) / sizeof(void *))];
 <% } %>
 
 
 <% _.forEach(data, function(name, index) { %>
   <% if (_.isArray(name)) { %>
-const AtomicString& k<%= name[0] %> = reinterpret_cast<AtomicString*>(&names_storage)[<%= index %>];
+thread_local const AtomicString& k<%= name[0] %> = reinterpret_cast<AtomicString*>(&names_storage)[<%= index %>];
   <% } else if (_.isObject(name)) { %>
-const AtomicString& k<%= name.name %> = reinterpret_cast<AtomicString*>(&names_storage)[<%= index %>];
+thread_local const AtomicString& k<%= name.name %> = reinterpret_cast<AtomicString*>(&names_storage)[<%= index %>];
   <% } else { %>
-const AtomicString& k<%= name %> = reinterpret_cast<AtomicString*>(&names_storage)[<%= index %>];<% } %>
+thread_local const AtomicString& k<%= name %> = reinterpret_cast<AtomicString*>(&names_storage)[<%= index %>];<% } %>
 <% }) %>
 
 <% if (deps && deps.html_attribute_names) { %>
   <% _.forEach(deps.html_attribute_names.data, function(name, index) { %>
-    const AtomicString& k<%= upperCamelCase(name) %>Attr = reinterpret_cast<AtomicString*>(&html_attribute_names_storage)[<%= index %>];
+   thread_local const AtomicString& k<%= upperCamelCase(name) %>Attr = reinterpret_cast<AtomicString*>(&html_attribute_names_storage)[<%= index %>];
   <% }) %>
 <% } %>
 
@@ -84,12 +84,14 @@ void Dispose(){
     AtomicString* atomic_string = reinterpret_cast<AtomicString*>(&names_storage) + i;
     atomic_string->~AtomicString();
   }
+  memset(names_storage, 0x00, sizeof(AtomicString) * kNamesCount);
 
   <% if (deps && deps.html_attribute_names) { %>
     for(size_t i = 0; i < kHtmlAttributeNamesCount; i ++) {
       AtomicString* atomic_string = reinterpret_cast<AtomicString*>(&html_attribute_names_storage) + i;
       atomic_string->~AtomicString();
     }
+    memset(html_attribute_names_storage, 0x00, sizeof(AtomicString) * kHtmlAttributeNamesCount);
   <% } %>
 };
