+++ bridge/webf_bridge.cc
@@ -8,7 +8,7 @@
 #include <thread>
 
 #include "bindings/qjs/native_string_utils.h"
-#include "core/dart_context.h"
+#include "core/dart_isolate_context.h"
 #include "core/page.h"
 #include "foundation/logging.h"
 #include "foundation/ui_command_buffer.h"
@@ -37,16 +37,9 @@
 #define SYSTEM_NAME "unknown"
 #endif
 
-thread_local webf::DartContext* dart_context_ = nullptr;
-
 void* initDartIsolateContext(uint64_t* dart_methods, int32_t dart_methods_len) {
-  if (dart_context_ == nullptr) {
-    dart_context_ = new webf::DartContext();
-  }
-
-  auto dart_isolate_context = std::make_unique<webf::DartIsolateContext>(dart_context_, dart_methods, dart_methods_len);
-  auto* ptr = dart_isolate_context.get();
-  dart_context_->AddIsolate(std::move(dart_isolate_context));
+  void* ptr = new webf::DartIsolateContext(dart_methods, dart_methods_len);
+  WEBF_LOG(VERBOSE) << "INIT DART ISOLATE:" << ptr;
   return ptr;
 }
 
@@ -160,13 +153,8 @@ int32_t profileModeEnabled() {
 // Callbacks when dart context object was finalized by Dart GC.
 static void finalize_dart_context(void* isolate_callback_data, void* peer) {
   auto* dart_isolate_context = (webf::DartIsolateContext*)peer;
-  auto* dart_context = dart_isolate_context->dartContext();
-  ((webf::DartIsolateContext*)peer)->dartContext()->RemoveIsolate(dart_isolate_context);
-  // Remove the whole dart context if there are no dart isolates alive.
-  if (dart_context->IsIsolateEmpty()) {
-    delete dart_context;
-    dart_context_ = nullptr;
-  }
+  WEBF_LOG(VERBOSE) << "FINALIZE DART ISOLATE CONTEXT: " << dart_isolate_context;
+  delete dart_isolate_context;
 }
 
 void init_dart_dynamic_linking(void* data) {
@@ -176,6 +164,6 @@ void init_dart_dynamic_linking(void* data) {
 }
 
 void register_dart_context_finalizer(Dart_Handle dart_handle, void* dart_isolate_context) {
-  Dart_NewFinalizableHandle_DL(dart_handle, reinterpret_cast<void*>(dart_isolate_context), sizeof(webf::DartContext),
+  Dart_NewFinalizableHandle_DL(dart_handle, reinterpret_cast<void*>(dart_isolate_context), sizeof(webf::DartIsolateContext),
                                finalize_dart_context);
 }
